Use font cache to fix mysterous fonts prevent transparent squares
- Fix https://github.com/chamons/ArenaGS/issues/183
- This one is a story. For the last fwe weeks, I've seen a few strange
cases of background squares not rendering correctly.
Well they are calling render, but nothing show up.
I came across it again today, and was able to pin down some things:
  - It happen after the first bit of text is copied onto the canvas
  - Removing the destroy makes it 'go away'
  - Many online sources don't seem to mention what's going on.
My best guess is that somehow the alpha channel is stored in the
surface, and destorying it before full rendering is somehow messing
with alpha channel info.

The 'hacky' solution is to make a circular ring of the last 32 font
Textures, and keep them alive until the ring comes around. This fixes
it for now, and holds a fix (small) consistent amount of memory around.